### 2017-05-17 #115 @njahn82
roadoi
roadoi interacts with the [oaDOI API](https://oadoi.org/), a simple web interface that links Digital Object Identifiers (DOIs) and open access full-texts disseminated by both open access journals and repositories.

```
Package: roadoi
Title: Find Free Versions of Scholarly Publications via the oaDOI Service
Version: 0.2
    person("Najko", "Jahn", role = c("aut", "cre"), email = "najko.jahn@gmail.com"))
Description: This web client interfaces oaDOI <https://oadoi.org>, a service finding
    free full-texts of academic papers by linking DOIs with open access journals and
    repositories. It provides unified access to various data sources for open access
    full-text links including Crossref, Bielefeld Academic Search Engine (BASE) and
    the Directory of Open Access Journals (DOAJ). API usage is free and no
    registration is required.
    httr,
    plyr,
    miniUI,
    shiny
    roxygen2 (>= 6.0.1),
    covr,
    rcrossref

```

<https://github.com/njahn82/roadoi>

-   Which categories does the package fall under from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit)? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

data retrieval

This package aims at academics who want to find openly available full-texts within R. More specifically, roadoi can be beneficial for text-mining because it retrieves links to full-texts, as well as for those studying scholarly communication. This package is also of practical use for scholarly communication librarians and analysts.

To my knowledge, there aren't any other R packages on CRAN wrapping the oaDOI-API. However, some R packages allow for searching and retrieving scholarly publications. One example, which is maintained by rOpenSci, is `fulltext`, a wrapper for searching freely available full-texts across many open access sources.

- [X] has continuous integration, including reporting of test coverage, using
services such as Travis CI, Coeveralls and/or CodeCov.

- [X] I agree to abide by [ROpenSci's Code of Conduct](https://github.com/ropensci/onboarding/blob/master/policies.md#code-of-conduct) during
the review process and in maintaining my package should it be accepted.

- [X] Do you intend for this package to go on CRAN? <- The package is already available via CRAN

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under and why:

I think this package falls best under the data retrieval category, because it supports the discovery of openly available scholarly publications. The oaDOI service itself claims to index more than 90 million articles covering sources like Crossref, a DOI minting agency, or the Bielefeld Academic Search Engine (BASE).

  - Ross Mounce - [rossmounce](https://github.com/rossmounce)
  - Tuija Sonkkila - [tts](https://github.com/tts)
