+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: roadoi
 Type: Package
 Title: Find Free Versions of Scholarly Publications via the oaDOI Service
-Version: 0.2
+Version: 0.2.9000
 Authors@R: c(
     person("Najko", "Jahn", role = c("aut", "cre"), email = "najko.jahn@gmail.com"))
 Description: This web client interfaces oaDOI <https://oadoi.org>, a service finding 
