### 2017-05-19 #115 @sckott
### Editor checks:

---

#### Editor comments

Currently seeking reviewers. It's a good fit and not overlapping.

* Played a bit with the package:
    * seems like `email` will take any string - maybe it's worth validating email addresses - seems like they don't server side either, though maybe they are just using it as a string and not validating it
* Ran `goodpractice::gp()`, not a lot to deal with. Test coverage is quite good. Could shorten line lengths to tidy things up a bit.

```r
It is good practice to

    R/oadoi_fetch.r:126:NA
    R/oadoi_fetch.r:127:NA
    R/oadoi_fetch.r:128:NA
    R/oadoi_fetch.r:129:NA
    R/oadoi_fetch.r:130:NA
    ... and 4 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your
    lines shorter than 80 characters

    R/oadoiAddins.R:24:1
```

---

Reviewers: xxx xxx
Due date: xxxxx
