### 2017-05-20 12dc2ad
adding tests
