+++ R/oadoi_fetch.r
@@ -118,7 +118,7 @@ oadoi_fetch_ <- function(doi = NULL, email) {
                     add_headers(
                       Accept = paste0("application/x.oadoi.",
                                       oadoi_api_version(), "+json")
-                    ), timeout(10), httr::verbose())
+                    ), timeout(10))
 
   # test for valid json
   if (httr::http_type(resp) != "application/json") {
