### 2017-05-20 #115 @njahn82
Thanks for your suggestions!  Email validation is now implemented and the long line is wrapped for better code readability.
