### 2017-05-24 0bd6d92
Adding example how to catch errors when calling oadoi_fetch() #9
