+++ README.md
@@ -23,7 +23,7 @@ information and full-text links from oaDOI.
 ```r
 roadoi::oadoi_fetch(dois = c("10.1038/ng.3260", "10.1093/nar/gkr1047"), 
                     email = "name@example.com")
-#> # A tibble: 2 × 20
+#> # A tibble: 2 x 20
 #>                                                              `_best_open_url`
 #>                                                                         <chr>
 #> 1 https://dash.harvard.edu/bitstream/handle/1/25290367/mallet%202015%20polyte
@@ -91,18 +91,19 @@ library(roadoi)
 roadoi::oadoi_fetch(dois = c("10.1186/s12864-016-2566-9",
                              "10.1016/j.cognition.2014.07.007"), 
                     email = "name@example.com")
-#> # A tibble: 2 × 20
-#>                                       `_best_open_url` `_closed_base_ids`
-#>                                                  <chr>             <list>
-#> 1             http://doi.org/10.1186/s12864-016-2566-9         <list [0]>
-#> 2 http://hdl.handle.net/11858/00-001M-0000-0024-2A9E-8          <chr [1]>
-#> # ... with 18 more variables: `_green_base_collections` <list>,
-#> #   `_open_base_ids` <list>, `_open_urls` <list>, doi <chr>,
-#> #   doi_resolver <chr>, evidence <chr>, found_green <lgl>,
-#> #   found_hybrid <lgl>, free_fulltext_url <chr>, is_boai_license <lgl>,
-#> #   is_free_to_read <lgl>, is_subscription_journal <lgl>, license <chr>,
-#> #   oa_color <chr>, oa_color_long <chr>,
-#> #   reported_noncompliant_copies <list>, url <chr>, year <int>
+#> # A tibble: 2 x 20
+#>                                                              `_best_open_url`
+#>                                                                         <chr>
+#> 1                                    http://doi.org/10.1186/s12864-016-2566-9
+#> 2 http://pubman.mpdl.mpg.de/pubman/item/escidoc:2070098/component/escidoc:207
+#> # ... with 19 more variables: `_closed_base_ids` <list>,
+#> #   `_green_base_collections` <list>, `_open_base_ids` <list>,
+#> #   `_open_urls` <list>, doi <chr>, doi_resolver <chr>, evidence <chr>,
+#> #   found_green <lgl>, found_hybrid <lgl>, free_fulltext_url <chr>,
+#> #   is_boai_license <lgl>, is_free_to_read <lgl>,
+#> #   is_subscription_journal <lgl>, license <chr>, oa_color <chr>,
+#> #   oa_color_long <chr>, reported_noncompliant_copies <list>, url <chr>,
+#> #   year <int>
 ```
 
 #### What's returned?
@@ -155,18 +156,42 @@ roadoi::oadoi_fetch(dois = c("10.1186/s12864-016-2566-9",
                     email = "name@example.com", 
                     .progress = "text")
 #>   |                                                                         |                                                                 |   0%  |                                                                         |================================                                 |  50%  |                                                                         |=================================================================| 100%
-#> # A tibble: 2 × 20
-#>                                       `_best_open_url` `_closed_base_ids`
-#>                                                  <chr>             <list>
-#> 1             http://doi.org/10.1186/s12864-016-2566-9         <list [0]>
-#> 2 http://hdl.handle.net/11858/00-001M-0000-0024-2A9E-8          <chr [1]>
-#> # ... with 18 more variables: `_green_base_collections` <list>,
-#> #   `_open_base_ids` <list>, `_open_urls` <list>, doi <chr>,
-#> #   doi_resolver <chr>, evidence <chr>, found_green <lgl>,
-#> #   found_hybrid <lgl>, free_fulltext_url <chr>, is_boai_license <lgl>,
[…]
