+++ DESCRIPTION
@@ -28,6 +28,7 @@ Suggests:
     covr,
     rmarkdown,
     lubridate,
-    rcrossref
+    rcrossref,
+    purrr
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
