### 2017-06-01 #115 @tts
---

This is a compact and useful package. For me personally timing was perfect because I had plans to add links to open access fulltext to my latest altmetrics Shiny app [aaltovirta](https://ttso.shinyapps.io/aaltovirta). Now links are in place, thanks to your work, @njahn82!

Although I do not really belong to the RStudio addin focus group, I can imagine that there are many people who welcome this functionality too.

`devtools::check()` and `devtools::test()` ran without errors on Ubuntu 16.04.2 LTS, R 3.2.3.

The RStudio addin works as expected.

* In `oadoi_fetch.R`, line 3: acccess --> access
* In `oadoi_fetch.R`, line 12: adress --> address
* In `oadoi_fetch.R`, lines 24-57: the following are returned by the function but missing from here: `_closed_base_ids`, `_closed_urls`, `oa_color_long`
* In `oadoi_fetch.R`, lines 24-57: columns are sorted by name which is fine. However, I wonder if the preceding column numbering is needed at all because as of now, it does not reflect the order of the result tibble. Or is it so that the `.Rd` command `\tabular` does not serve very well here?
* In `oadoi_fetch.R`, line 34: green_base_collections --> _green_base_collections
* In vignette `intro.Rmd`, line 27: json --> JSON
