### 2017-06-12 njahn82/roadoi#10 @rossmounce
Error : object ‘textAreaInput’ is not exported by 'namespace:shiny'
Apologies, for my late start on getting around to review your package!

I had trouble installing from CRAN, I'll move on to trying the development version after this.

```
> install.packages("roadoi")
Installing package into ‘/home/ross/R/x86_64-pc-linux-gnu-library/3.2’
(as ‘lib’ is unspecified)
--2017-06-12 07:29:55--  https://cran.rstudio.com/src/contrib/roadoi_0.2.tar.gz
Resolving cran.rstudio.com (cran.rstudio.com)... 52.85.67.12
Connecting to cran.rstudio.com (cran.rstudio.com)|52.85.67.12|:443... connected.
HTTP request sent, awaiting response... 200 OK
Length: 559051 (546K) [application/x-gzip]
Saving to: ‘/tmp/RtmpAkRkEv/downloaded_packages/roadoi_0.2.tar.gz’

     0K .......... .......... .......... .......... ..........  9%  451K 1s
    50K .......... .......... .......... .......... .......... 18%  216K 2s
   100K .......... .......... .......... .......... .......... 27% 3.83M 1s
   150K .......... .......... .......... .......... .......... 36% 2.76M 1s
   200K .......... .......... .......... .......... .......... 45%  172K 1s
   250K .......... .......... .......... .......... .......... 54% 3.07M 1s
   300K .......... .......... .......... .......... .......... 64% 3.60M 0s
   350K .......... .......... .......... .......... .......... 73% 2.58M 0s
   400K .......... .......... .......... .......... .......... 82% 4.33M 0s
   450K .......... .......... .......... .......... .......... 91% 2.76M 0s
   500K .......... .......... .......... .......... .....     100% 4.61M=0.8s

2017-06-12 07:29:56 (728 KB/s) - ‘/tmp/RtmpAkRkEv/downloaded_packages/roadoi_0.2.tar.gz’ saved [559051/559051]

* installing *source* package ‘roadoi’ ...
** package ‘roadoi’ successfully unpacked and MD5 sums checked
** R
** inst
** preparing package for lazy loading
Error : object ‘textAreaInput’ is not exported by 'namespace:shiny'
ERROR: lazy loading failed for package ‘roadoi’
* removing ‘/home/ross/R/x86_64-pc-linux-gnu-library/3.2/roadoi’
Warning in install.packages :
  installation of package ‘roadoi’ had non-zero exit status

The downloaded source packages are in
	‘/tmp/RtmpAkRkEv/downloaded_packages’
> sessionInfo()
R version 3.2.5 (2016-04-14)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 16.04.2 LTS

locale:
 [1] LC_CTYPE=en_GB.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_GB.UTF-8        LC_COLLATE=en_GB.UTF-8
 [5] LC_MONETARY=en_GB.UTF-8    LC_MESSAGES=en_GB.UTF-8
 [7] LC_PAPER=en_GB.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_GB.UTF-8 LC_IDENTIFICATION=C

attached base packages:

[1] tools_3.2.5
```
