### 2017-06-12 #115 @rossmounce
@sckott thanks for the heads up. I've started to look into the package but I'll probably need a few more days (perhaps?).
