### 2017-06-12 ropensci/roadoi#10 @njahn82
Not good. Maybe you are running an outdated version of shiny. Try installing the most recent version from CRAN.

```r
install.packages("shiny")
```

If this trick helps to solve the problem, then I'll have to be more explicit about which `shiny` version to use in `DESCRIPTION` in order to avoid these errors.
