### 2017-06-12 njahn82/roadoi#11 @rossmounce
what is "blue" open access?
I notice from the README that you have green, gold, and blue options for `oa_color`.

The oaDOI API seemingly only returns gold & green(?): https://oadoi.org/api

I'm guessing "blue" OA might be equivalent to when `is_boai_license` is `TRUE` but I couldn't find this explicitly stated anywhere. So at the very least some clarification is needed.

Personally I'm not a fan of the terms 'gold' and 'green' - they are oft misunderstood and aren't very fitting words for things they describe ('gold' often implies expensive, but it doesn't have to be). But gold and green are popular terms and thus are probably here to stay whether I like it or not.

The same can't be said for blue OA. I think very people in the world know what blue OA is and thus I'd urge you not to use that term in this package. The functionality is great - keep that, it's good to know what subset of gold OA is BOAI-compliant. Just rename it BOAI-compliant OA or something like that?
