### 2017-06-12 ropensci/roadoi#10 @njahn82
Because I am not running Ubuntu, I used Travis CI to test the CRAN version again.

https://travis-ci.org/njahn82/roadoi/builds/233283484

Installation should work with

```
shiny         1.0.3    2017-04-26 cran (@1.0.3)
```

To check the versions of your installed packages, you can use

```r
devtools::session_info(installed.packages()[, "Package"])
```
