### 2017-06-13 ropensci/roadoi#11 @njahn82
Thanks for bringing this up. I also wondered what is meant by blue OA. I found the following example

https://api.oadoi.org/10.1016/j.bpj.2012.11.2487

```
{
  "results": [
    {
      "_best_open_url": "http://doi.org/10.1016/j.bpj.2012.11.2487",
      "_closed_base_ids": [],
      "_closed_urls": [],
      "_green_base_collections": [],
      "_open_base_ids": [
        "crelsevierbv:10.1016/j.bpj.2012.11.2487"
      ],
      "_open_urls": [
        "http://doi.org/10.1016/j.bpj.2012.11.2487",
        "http://www.cell.com/article/S0006349512037332/pdf"
      ],
      "_title": "Surface Immobilization of Cardiac Thin Filaments",
      "doi": "10.1016/j.bpj.2012.11.2487",
      "doi_resolver": "crossref",
      "evidence": "hybrid (via crossref license)",
      "found_green": false,
      "found_hybrid": true,
      "free_fulltext_url": "http://doi.org/10.1016/j.bpj.2012.11.2487",
      "is_boai_license": false,
      "is_free_to_read": true,
      "is_subscription_journal": true,
      "license": "elsevier-specific: oa user license",
      "oa_color": "blue",
      "oa_color_long": "gold_hybrid",
      "reported_noncompliant_copies": [],
      "url": "http://doi.org/10.1016/j.bpj.2012.11.2487",
      "year": 2013
    }
  ]
}
```

My guess, it is somehow related to hybrid, but I am not sure. Pinging @impactstory What is meant by `blue OA`?
