### 2017-06-13 ropensci/roadoi#10 @njahn82
@rossmounce Did the installation work after updating the shiny package?
