### 2017-06-14 d669da4
adding lintr unit test example from @maelle rugbcn_20170613_improve_pkg talk. Had to exclude Shiny addin though.
