### 2017-06-14 900669e
improve linting d669da4cad9bc447ad945ff929f8d60ef3b1b974
