+++ R/oadoiAddins.R
@@ -14,13 +14,12 @@
 #' }
 #'
 #' @export
-
-# nolint start
-roadoi_addin <- function() { # nocov start
+roadoi_addin <- function() {
+  # nocov start
   # create user interface like the one rcrossref provides
-  ui <- miniPage(
-    gadgetTitleBar("Find freely available full-text via oaDOI.org"),
-    miniContentPanel(
+  ui <- miniUI::miniPage(
+    miniUI::gadgetTitleBar("Find freely available full-text via oaDOI.org"),
+    miniUI::miniContentPanel(
       tags$h4("Find fulltexts for scholarly articles"),
       tags$p(
         "If you have DOIs (Digital Object Identifier) for several articles
@@ -28,19 +27,19 @@ roadoi_addin <- function() { # nocov start
         in the text box below. Please note that only the first ten DOIs will
         be fetched."
       ),
-      textAreaInput(
+      shiny::textAreaInput(
         inputId = "text",
         label = "DOIs (line separated):",
         value = "10.1007/s13752-012-0049-z\n10.1098/rsta.2016.0122",
         height = 200
       ),
-      actionButton(inputId = "submit", "Run!"),
-      tableOutput("table")
+      shiny::actionButton(inputId = "submit", "Run!"),
+      shiny::tableOutput("table")
     )
   )
   # here's the server-side R code that will be executed to find OA copies
   server <- function(input, output) {
-    observeEvent(input$submit, {
+    shiny::observeEvent(input$submit, {
       my_input <- reactive({
         # avoid warnings about "no visible binding for global variable"
         `Free fulltext link` <- NULL
@@ -64,7 +63,7 @@ roadoi_addin <- function() { # nocov start
           ))
       })
       # output as table
-      output$table <- renderTable(
+      output$table <- shiny::renderTable(
         my_input(),
         sanitize.text.function = function(x)
           x
@@ -73,15 +72,16 @@ roadoi_addin <- function() { # nocov start
   }
 
   # finally, define where and how the gadget is displayed
-  viewer <- dialogViewer("Find free full-texts",
+  viewer <- shiny::dialogViewer("Find free full-texts",
                          width = 800, height = 800)
-  runGadget(ui, server, viewer = viewer)
+  shiny::runGadget(ui, server, viewer = viewer)
 }
 
 # helper to make links clickable
+# nolint start
 create_link <- function(x) {
   paste0('<a href="', x, '" target="_blank">', x, '</a>')
 }
+# nolint end
 
 # nocov end
-# nolint end
