### 2017-06-15 ropensci/roadoi#10 @rossmounce
@njahn82 yes thanks,
running
```
install.packages('shiny')
```
to install `shiny` 1.0.3 before installing `roadoi` via CRAN worked.
