### 2017-06-15 #115 @rossmounce
---

This is an excellent wrapper for the R universe, for an excellent API and and an excellent set of data resources beneath it all. It's really impressive to see that with tools like this, open access is really starting to come together and be more of a seemless and universal user-experience.

As a new OA grants manager, I'm hugely interested in tools that'll make my life easier/quicker/more accurate when attempting to evaluate the open access compliance of our grantees work. I will be sure to use this package in my compliance workflow.

Furthermore, I might discuss this with other open access administrators on the next open research funders group (ORFG) call. There are a lot of smaller nonprofits and charities out there that will need a free, open-source software solution to help them with compliance checking of outputs so I'm really glad this package exists.

Once I get to use this package more, I'm sure I'll have more comments/issues but for the moment I just want to get this review done so it isn't a blocker.

devtools::check() and devtools::test() ran without errors on Ubuntu 16.04.2 LTS, R 3.2.5. When I get permission to install Rstudio on my 'work' machine, I will let you how the tests run under Win10.

I did not try the RStudio addin.

In `vignettes/intro.Rmd`, line 68: when you are tired to type in your email --> when you are too tired to type in your email

Installation via devtools & github was smooth, but I ran into a small problem when trying to install via CRAN which I reported as an issue: https://github.com/njahn82/roadoi/issues/10  Perhaps there should be a dependencies/pre-requisites section in the README so that new users might be made more aware of what they need before attempting to install `roadoi` ?

I was intrigued by what "blue OA" was, and I still have no concrete idea of what it is. I reported this as an issue here and am waiting to hear back from Impactstory about it: https://github.com/njahn82/roadoi/issues/11
