### 2017-06-15 #115 @sckott
thanks for your @rossmounce !

@njahn82 continue discussion here - can also ping editors with any questions
