### 2017-06-16 ropensci/roadoi#11 @rossmounce
I've had a bit more of a dig around the unpaywall site and I think the explanation is there:

"Blue tab for articles available on the current page, but lacking license information (often that's because you're browsing from behind the paywall)"

http://unpaywall.org/faq

In practise this appears to mark articles where there is a moving wall of delayed free availability, say 6-months, 12-months or 24-months. But I guess there might be circumstances outside of the moving-wall scenario that also fall into apparent "blue".
