### 2017-06-16 ropensci/roadoi#11 @njahn82
Good to know, will add it to documentation together with a note that oa colour blue is quite unusual.
