### 2017-06-23 e42bdea
answering @tts onboarding comments
