+++ man/oadoi_fetch.Rd
@@ -13,7 +13,7 @@ A rate limit of 100k requests per day is suggested. If you need to access
 more data, use the data dump \url{https://oadoi.org/api#dataset} instead.}
 
 \item{email}{character vector, required! It is strongly encourage to tell
-oaDOI your email adress, so that they can track usage and notify you
+oaDOI your email address, so that they can track usage and notify you
 when something breaks. Set email address in your `.Rprofile` file with
 the option `roadoi_email` \code{options(roadoi_email = "name@example.com")}.}
 
@@ -24,58 +24,65 @@ of each. By default, no progress bar is displayed.}
 }
 \value{
 The result is a tibble with each row representing a publication and
-and the following columns.
+  and the following columns. Here are some fields returned by the API.
+  However, other fields can be added by oaDOI over time.
 
-\tabular{rll}{
- [,1] \tab `_best_open_url`   \tab link to free full-text \cr
- [,2] \tab doi                \tab DOI \cr
- [,3] \tab doi_resolver       \tab DOI agency \cr
- [,4] \tab evidence           \tab A phrase summarizing the step of the
- open access detection process where the full-text links were found. \cr
- [,5] \tab found_green        \tab logical indicating whether a self-archived
- copy in a repository was found \cr
- [,6] \tab found_hybrid       \tab logical indicating whether an open access
- article was published in a toll-access journal \cr
- [,7] \tab free_fulltext_url  \tab URL where the free version was found \cr
- [,8] \tab green_base_collections \tab internal collection ID from the
+
+\tabular{ll}{
+ \code{`_best_open_url`} \tab link to free full-text \cr
+ \code{`_closed_base_ids`}   \tab internal ids for closed access copies \cr
+ returned from the Bielefeld Academic Search Engine (BASE) \cr
+ \code{`_closed_urls`}   \tab links to closed access copies \cr
+ \code{`_green_base_collections`} \tab internal collection ID from the
  Bielefeld Academic Search Engine (BASE) \cr
- [,9] \tab is_boai_license    \tab TRUE whenever the license indications are
+ \code{`_open_base_ids`} \tab  ids of oai metadata records with open access
+ full-text links collected by the Bielefeld Academic Search Engine (BASE) \cr
+ \code{`_open_urls`}     \tab full-text url \cr
+ \code{`_title`}         \tab title of the work \cr
+ \code{doi}              \tab DOI \cr
+ \code{doi_resolver}     \tab DOI agency \cr
+ \code{evidence}         \tab A phrase summarizing the step of the
+ open access detection process where the full-text links were found. \cr
+ \code{found_green}      \tab logical indicating whether a self-archived
+ copy in a repository was found \cr
+ \code{found_hybrid}     \tab logical indicating whether an open access
+ article was published in a toll-access journal \cr
+ \code{free_fulltext_url}\tab URL where the free version was found \cr
+ \code{is_boai_license}  \tab TRUE whenever the license indications are
  Creative Commons - Attribution (CC BY), Creative Commons CC - Universal(CC 0))
  or Public Domain. These permissive licenses comply with the
  highly-regarded Budapest Open Access Initiative (BOAI) definition of
  open access \cr
- [,10] \tab is_free_to_read  \tab TRUE if freely available full-text
+ \code{is_free_to_read}  \tab TRUE if freely available full-text
  was found \cr
- [,11] \tab is_subscription_journal \tab TRUE if article is published in
+ \code{is_subscription_journal} \tab TRUE if article is published in
  toll-access journal \cr
- [,12] \tab license          \tab Contains the name of the Creative
[…]
