### 2017-06-23 d8ab52a
answering @rossmounce suggestions and comments
