+++ vignettes/intro.Rmd
@@ -63,9 +63,11 @@ According to the [oaDOI.org API specification](https://oadoi.org/api), the follo
 * `url`: the canonical DOI URL
 * `year`: year of publication
 
+Note that fields to be returned might change according to the [oaDOI.org API specs](https://oadoi.org/api)
+
 #### Any API restrictions?
 
-There are no API restrictions. However, providing your email address when using this client is required by oaDOI.org. Set email address in your `.Rprofile` file with the option `roadoi_email` when you are tired to type in your email address every time you want to call oadDOI.
+There are no API restrictions. However, providing your email address when using this client is required by oaDOI.org. Set email address in your `.Rprofile` file with the option `roadoi_email` when you are too tired to type in your email address every time you want to call oadDOI.
 
 ```r
 options(roadoi_email = "name@example.com")
