+++ README.md
@@ -23,18 +23,19 @@ information and full-text links from oaDOI.
 ```r
 roadoi::oadoi_fetch(dois = c("10.1038/ng.3260", "10.1093/nar/gkr1047"), 
                     email = "name@example.com")
-#> # A tibble: 2 x 20
+#> # A tibble: 2 x 23
 #>                                                              `_best_open_url`
 #>                                                                         <chr>
 #> 1 https://dash.harvard.edu/bitstream/handle/1/25290367/mallet%202015%20polyte
 #> 2                                          http://doi.org/10.1093/nar/gkr1047
-#> # ... with 19 more variables: `_closed_base_ids` <list>,
-#> #   `_green_base_collections` <list>, `_open_base_ids` <list>,
-#> #   `_open_urls` <list>, doi <chr>, doi_resolver <chr>, evidence <chr>,
-#> #   found_green <lgl>, found_hybrid <lgl>, free_fulltext_url <chr>,
-#> #   is_boai_license <lgl>, is_free_to_read <lgl>,
-#> #   is_subscription_journal <lgl>, license <chr>, oa_color <chr>,
-#> #   oa_color_long <chr>, reported_noncompliant_copies <list>, url <chr>,
+#> # ... with 22 more variables: `_closed_base_ids` <list>,
+#> #   `_closed_urls` <list>, `_green_base_collections` <list>,
+#> #   `_open_base_ids` <list>, `_open_urls` <list>, `_title` <chr>,
+#> #   doi <chr>, doi_resolver <chr>, evidence <chr>, found_green <lgl>,
+#> #   found_hybrid <lgl>, free_fulltext_url <chr>, is_boai_license <lgl>,
+#> #   is_free_to_read <lgl>, is_subscription_journal <lgl>, license <chr>,
+#> #   oa_color <chr>, oa_color_long <chr>,
+#> #   reported_noncompliant_copies <list>, url <chr>, version <lgl>,
 #> #   year <int>
 ```
 
@@ -91,18 +92,19 @@ library(roadoi)
 roadoi::oadoi_fetch(dois = c("10.1186/s12864-016-2566-9",
                              "10.1016/j.cognition.2014.07.007"), 
                     email = "name@example.com")
-#> # A tibble: 2 x 20
+#> # A tibble: 2 x 23
 #>                                                              `_best_open_url`
 #>                                                                         <chr>
 #> 1                                    http://doi.org/10.1186/s12864-016-2566-9
 #> 2 http://pubman.mpdl.mpg.de/pubman/item/escidoc:2070098/component/escidoc:207
-#> # ... with 19 more variables: `_closed_base_ids` <list>,
-#> #   `_green_base_collections` <list>, `_open_base_ids` <list>,
-#> #   `_open_urls` <list>, doi <chr>, doi_resolver <chr>, evidence <chr>,
-#> #   found_green <lgl>, found_hybrid <lgl>, free_fulltext_url <chr>,
-#> #   is_boai_license <lgl>, is_free_to_read <lgl>,
-#> #   is_subscription_journal <lgl>, license <chr>, oa_color <chr>,
-#> #   oa_color_long <chr>, reported_noncompliant_copies <list>, url <chr>,
+#> # ... with 22 more variables: `_closed_base_ids` <list>,
+#> #   `_closed_urls` <list>, `_green_base_collections` <list>,
+#> #   `_open_base_ids` <list>, `_open_urls` <list>, `_title` <chr>,
+#> #   doi <chr>, doi_resolver <chr>, evidence <chr>, found_green <lgl>,
+#> #   found_hybrid <lgl>, free_fulltext_url <chr>, is_boai_license <lgl>,
[…]
