### 2017-06-23 #115 @njahn82
I have now addressed  the reviewer concerns, thanks again to @tts and @rossmounce for taking the time to review the package and for your very helpful suggestions. I am so glad to read that you not only see potential use of this package for your work, but that you have already used it.

All changes made to the source code can be found here:
https://github.com/njahn82/roadoi/pull/14

@tts comments and suggestions:

> In oadoi_fetch.R, line 3: acccess --> access

Fixed.

>In oadoi_fetch.R, line 12: adress --> address

Fixed.

> In oadoi_fetch.R, lines 24-57: the following are returned by the function but missing from here: _closed_base_ids, _closed_urls, oa_color_long

Added. I also added a comment that, according to the API documentation, return fields are not fixed and some of them may be added or removed by oaDOI

> In oadoi_fetch.R, lines 24-57: columns are sorted by name which is fine. However, I wonder if the preceding column numbering is needed at all because as of now, it does not reflect the order of the result tibble. Or is it so that the .Rd command \tabular does not serve very well here?

Agreed. I removed the numbering in the table.

> In oadoi_fetch.R, line 34: green_base_collections --> _green_base_collections

Fixed.

> In vignette intro.Rmd, line 27: json --> JSON

Fixed.

@rossmounce comments and suggestions:

> In vignettes/intro.Rmd, line 68: when you are tired to type in your email --> when you are too tired to type in your email

Fixed

> Installation via devtools & github was smooth, but I ran into a small problem when trying to install via CRAN which I reported as an issue: njahn82/roadoi#10 Perhaps there should be a dependencies/pre-requisites section in the README so that new users might be made more aware of what they need before attempting to install roadoi ?

Thank you for making me aware of this dependency. I now added the required version of the `shiny` package to the DESCRIPTION file as dependency.

> I was intrigued by what "blue OA" was, and I still have no concrete idea of what it is. I reported this as an issue and am waiting to hear back from Impactstory about it: njahn82/roadoi#11

Agreed. The term `blue open access` is rather uncommon and the meaning is not clear. I am still waiting for an answer from the Impactstory team and will add it to the documentation.

Because of your valuable input, I would like to add you as contributors to the `DESCRIPTION`. Are you ok with that?
