### 2017-06-24 #115 @noamross
@njahn82 re: adding contributors, we have a little experiment on that front.  So as not to clutter up this review thread, can you take a look at https://github.com/ropensci/onboarding-meta/issues/1 ?
