### 2017-06-27 #115 @rossmounce
I'd be delighted to be listed as a contributor. Thanks @njahn82
