### 2017-06-27 #115 @sckott
Thanks for making changes @njahn82

@tts and @rossmounce - Are you happy with the changes/fixes?  Any other concerns/thoughts you have?
