### 2017-06-28 #115 @tts
Quite happy with fixes from my part.

One thing I didn't check earlier was the behaviour of the `Done` button in the RStudio addin. At the moment, it doesn't respond to a click. AFAIK you need to add a listener to `input$done` Something like

```
shiny::observeEvent(input$done, {
      stopApp()
    })
```
as [explained here](http://rstudio.github.io/rstudioaddins/#gadget-server)

From my current (default) addins, [Add Crossref Citations](https://github.com/ropensci/rcrossref/blob/master/R/crAddins.R) wraps `stopApp` inside `invisible` but I'm not sure if that's necessary.
