### 2017-06-28 90f7d85
bugfix stopping shiny addin
