### 2017-06-28 #115 @njahn82
Very helpful @tts, I fixed the bug as suggested.
