### 2017-07-14 #115 @sckott
Approved! thank for your submission @njahn82 !

* Please transfer the package  to `ropensci`- I've invited you to a team within our ropensci organization. After you accept you should be able to move it.
* Make sure to
    * add rOpenSci footer to README
* Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ?  If so, i'll get you the details on that
