+++ README.md
@@ -3,11 +3,12 @@
 
 
 
-[![Build Status](https://travis-ci.org/njahn82/roadoi.svg?branch=master)](https://travis-ci.org/njahn82/roadoi)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/njahn82/roadoi?branch=master&svg=true)](https://ci.appveyor.com/project/njahn82/roadoi)
-[![codecov.io](https://codecov.io/github/njahn82/roadoi/coverage.svg?branch=master)](https://codecov.io/github/njahn82/roadoi?branch=master)
+[![Build Status](https://travis-ci.org/ropensci/roadoi.svg?branch=master)](https://travis-ci.org/ropensci/roadoi)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/roadoi?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/roadoi)
+[![codecov.io](https://codecov.io/github/ropensci/roadoi/coverage.svg?branch=master)](https://codecov.io/github/ropensci/roadoi?branch=master)
 [![cran version](http://www.r-pkg.org/badges/version/roadoi)](https://cran.r-project.org/package=roadoi)
 [![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/roadoi)](https://github.com/metacran/cranlogs.app)
+[![review](https://ropensci.org/badges/115_status.svg)](https://github.com/ropensci/onboarding/issues/115)
 
 roadoi interacts with the [oaDOI API](http://oadoi.org/), a simple interface which links DOIs 
 and open access versions of scholarly works. oaDOI powers [unpaywall](http://unpaywall.org/).
@@ -23,15 +24,15 @@ information and full-text links from oaDOI.
 ```r
 roadoi::oadoi_fetch(dois = c("10.1038/ng.3260", "10.1093/nar/gkr1047"), 
                     email = "name@example.com")
-#> # A tibble: 2 x 23
+#> # A tibble: 2 x 22
 #>                                                              `_best_open_url`
 #>                                                                         <chr>
 #> 1 https://dash.harvard.edu/bitstream/handle/1/25290367/mallet%202015%20polyte
 #> 2                                          http://doi.org/10.1093/nar/gkr1047
-#> # ... with 22 more variables: `_closed_base_ids` <list>,
+#> # ... with 21 more variables: `_closed_base_ids` <list>,
 #> #   `_closed_urls` <list>, `_green_base_collections` <list>,
-#> #   `_open_base_ids` <list>, `_open_urls` <list>, `_title` <chr>,
-#> #   doi <chr>, doi_resolver <chr>, evidence <chr>, found_green <lgl>,
+#> #   `_open_base_ids` <list>, `_open_urls` <list>, doi <chr>,
+#> #   doi_resolver <chr>, evidence <chr>, found_green <lgl>,
 #> #   found_hybrid <lgl>, free_fulltext_url <chr>, is_boai_license <lgl>,
 #> #   is_free_to_read <lgl>, is_subscription_journal <lgl>, license <chr>,
 #> #   oa_color <chr>, oa_color_long <chr>,
@@ -61,7 +62,7 @@ To install the development version, use the [devtools package](https://cran.r-pr
 
 
 ```r
-devtools::install_github("njahn82/roadoi")
+devtools::install_github("ropensci/roadoi")
 library(roadoi)
 ```
 
@@ -92,15 +93,15 @@ library(roadoi)
 roadoi::oadoi_fetch(dois = c("10.1186/s12864-016-2566-9",
                              "10.1016/j.cognition.2014.07.007"), 
                     email = "name@example.com")
-#> # A tibble: 2 x 23
+#> # A tibble: 2 x 22
[…]

[…]
