### 2017-07-16 #115 @njahn82
Great, thank you very much for accepting this package. I am very grateful about the reviews, which were very helpful to improve this package!

I transferred the repo as described. After adding the "peer-review" badge, I get the following warning when calling `R CMD Check --as-cran`

```
* checking top-level files ... WARNING
Conversion of ‘README.md’ failed:
HttpExceptionRequest Request {
  host                 = "ropensci.org"
  port                 = 443
  secure               = True
  requestHeaders       = []
  queryString          = ""
  method               = "GET"
  proxy                = Nothing
  rawBody              = False
  redirectCount        = 10
  responseTimeout      = ResponseTimeoutDefault
  requestVersion       = HTTP/1.1
}
 (InternalException (HandshakeFailed Error_EOF))
```

Can you help @sckott ?
