### 2017-07-16 #115 @njahn82
My fault, I was running pandoc 1.18, which caused this warning. Upgrading to the newest pandoc release helps.
