### 2017-07-16 #115 @njahn82
> Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ? either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development (https://ropensci.org/blog/). If so, i'll get you the details on that

Sounds good. Will prepare a draft and will send it to you by mail within the next two weeks.
