### 2017-07-17 d80828d
prepare CRAN re-submission because of non-canonical CRAN link and because the reviewer role was not accepted
