+++ README.md
@@ -179,7 +179,7 @@ roadoi::oadoi_fetch(dois = c("10.1186/s12864-016-2566-9",
 
 #### Catching errors
 
-oaDOI is a reliable API. However, this client follows [Hadley Wickham's Best practices for writing an API package](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html) and throws an error when API does not return valid JSON or is not available. To catch these errors, you may want to use [plyr's `failwith()`](https://www.rdocumentation.org/packages/plyr/versions/1.8.4/topics/failwith) function
+oaDOI is a reliable API. However, this client follows [Hadley Wickham's Best practices for writing an API package](https://CRAN.R-project.org/package=httr/vignettes/api-packages.html) and throws an error when API does not return valid JSON or is not available. To catch these errors, you may want to use [plyr's `failwith()`](https://www.rdocumentation.org/packages/plyr/versions/1.8.4/topics/failwith) function
 
 
 ```r
@@ -219,26 +219,26 @@ random_dois <- rcrossref::cr_r(sample = 100) %>%
   .$data
 random_dois
 #> # A tibble: 100 x 35
-#>       alternative.id
-#>                <chr>
-#>  1                  
-#>  2                  
-#>  3                  
-#>  4                  
-#>  5                47
-#>  6                  
-#>  7 S095758201730071X
-#>  8 S0022460X12005676
-#>  9                  
-#> 10 10.1021/ic702144n
+#>                                     alternative.id
+#>                                              <chr>
+#>  1                                                
+#>  2                                         5088422
+#>  3                                                
+#>  4              10.1111/j.1478-1913.1971.tb03042.x
+#>  5                                                
+#>  6                             10.1021/j100789a044
+#>  7                  10.1080/00107530.1989.10746306
+#>  8                                                
+#>  9                                            2758
+#> 10 10.1201/9781420063561.ch8,10.1201/9781420063561
 #> # ... with 90 more rows, and 34 more variables: container.title <chr>,
 #> #   created <chr>, deposited <chr>, DOI <chr>, funder <list>,
-#> #   indexed <chr>, ISBN <chr>, ISSN <chr>, issue <chr>, issued <chr>,
-#> #   link <list>, member <chr>, page <chr>, prefix <chr>, publisher <chr>,
+#> #   indexed <chr>, ISBN <chr>, ISSN <chr>, issued <chr>, link <list>,
+#> #   member <chr>, page <chr>, prefix <chr>, publisher <chr>,
 #> #   reference.count <chr>, score <chr>, source <chr>, subject <chr>,
-#> #   title <chr>, type <chr>, URL <chr>, volume <chr>, assertion <list>,
-#> #   author <list>, `clinical-trial-number` <list>, license_date <chr>,
-#> #   license_URL <chr>, license_delay.in.days <chr>,
+#> #   title <chr>, type <chr>, URL <chr>, assertion <list>, author <list>,
+#> #   `clinical-trial-number` <list>, issue <chr>, volume <chr>,
+#> #   license_date <chr>, license_URL <chr>, license_delay.in.days <chr>,
 #> #   license_content.version <chr>, update.policy <chr>, subtitle <chr>,
 #> #   abstract <chr>, archive <chr>
 ```
@@ -254,20 +254,20 @@ random_dois %>%
   group_by(issued) %>%
   summarize(pubs = n()) %>%
   arrange(desc(pubs))
-#> # A tibble: 41 x 2
+#> # A tibble: 52 x 2
 #>    issued  pubs
 #>     <dbl> <int>
-#>  1     NA     9
-#>  2   2016     8
-#>  3   2013     6
-#>  4   2015     6
-#>  5   2012     5
-#>  6   1993     4
[…]
