+++ DESCRIPTION
@@ -4,10 +4,10 @@ Title: Find Free Versions of Scholarly Publications via the oaDOI Service
 Version: 0.3
 Authors@R: c(
     person("Najko", "Jahn", role = c("aut", "cre"), email = "najko.jahn@gmail.com"),
-    person("Tuija", "Sonkkila", role = c("rev"), comment = "Tuija Sonkkila 
+    person("Tuija", "Sonkkila", role = c("ctb"), comment = "Tuija Sonkkila 
     reviewed the package for rOpenSci, 
     see https://github.com/ropensci/onboarding/issues/115"),
-    person("Ross Mounce", role = c("rev"), comment = "Ross Mounce
+    person("Ross Mounce", role = c("ctb"), comment = "Ross Mounce
     reviewed the package for rOpenSci, 
     see https://github.com/ropensci/onboarding/issues/115")
     )
