### 2017-07-17 #115 @njahn82
Is on CRAN now!
